Fix/improve <input> in <label>
According to the HTML specification, some elements such as <input> are void elements, which means they do not have any content and do not need a closing tag. However, you can optionally use a self-closing tag by adding a slash before the closing bracket, such as <input type="text" />. This is useful if you want your markup to be valid XML as well.

If you want to add a label to your input element, you should use a separate <label> element and associate it with the input element using the id and for attributes. For example:

<input id="my_id" type="radio" name="radio_name">
<label for="my_id">Radio Label</label>
This way, the label will be displayed next to the input element and will also be clickable.